+++ src/contentEntries.js
@@ -4,10 +4,12 @@ import { grab } from './apiHelper';
 import { getValidatedOptions } from './options';
 
 const getRequestUrl = (url, path, params = {}) => {
-  const { id, preview, at, typeId, offset, size } = params;
+  const { id, typeId, alias, preview, at, offset, size } = params;
   const qsParams = {};
   // The id array must be turned into CSV
   if (id && id.length) { qsParams.id = `${id.join(',')}`; }
+  // The alias array must be turned into CSV
+  if (alias && alias.length) { qsParams.alias = `${alias.join(',')}`; }
   // preview is only useful when true
   if (preview) { qsParams.preview = true; }
   // at is either a string, or a method with toISOString (like a Date) that we use for formatting
@@ -28,7 +30,8 @@ const validateAndRequest = (opts, path, params) =>
     return grab(requestUrl, options);
   });
 
-// params should contain any/several of { id, preview, at, typeId, offset, size }
+// params should contain any/several of { id, alias, preview, at, typeId, offset, size }
+// do not use id, alias or typeId at the same time - behaviour would be ungaranteed
 export const getEntries = (opts, params) =>
   validateAndRequest(opts, 'content/entries', params);